+++ app/Http/Controllers/SkillController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\SkillStoreRequest;
-use App\Http\Requests\SkillUpdateRequest;
+use App\Http\Requests\SkillRequest;
 use App\Models\Skill;
+use App\Repositories\Languages;
 use App\Repositories\Skills;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Console\Application as ConsoleApplication;
 use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Contracts\Support\Renderable;
@@ -28,15 +29,24 @@ class SkillController extends AdminController
     protected Skills $skills;
 
     /**
+     * Languages repository instance.
+     *
+     * @param Languages $languages
+     */
+    protected Languages $languages;
+
+    /**
      * Class constructor.
      *
      * @return void
      */
-    public function __construct(Request $request, Skills $skills)
+    public function __construct(Request $request, Skills $skills, Languages $languages)
     {
         parent::__construct($request);
 
         $this->skills = $skills;
+
+        $this->languages = $languages;
     }
 
     /**
@@ -46,12 +56,14 @@ class SkillController extends AdminController
      * Returns the skills view.
      *
      * @return ConsoleApplication|FoundationApplication|View|Factory|Response|ResponseFactory
-     * @throws BindingResolutionException
      */
     public function index(): ConsoleApplication|FoundationApplication|View|Factory|Response|ResponseFactory
     {
         $skills = $this->skills->listing($this->skills->newQuery()->orderBy('order'));
-        return view('admin.skills.index', compact('skills'));
+
+        $languages = $this->languages->newQuery()->orderByLocale()->get();
+
+        return view('admin.skills.index', compact('skills','languages'));
     }
 
     /**
@@ -61,10 +73,12 @@ class SkillController extends AdminController
      * Returns the skill modal stream view for create.
      *
      * @return RedirectResponse|Response|ResponseFactory
-     * @throws BindingResolutionException
+     * @throws BindingResolutionException|AuthorizationException
      */
     public function create(): RedirectResponse|Response|ResponseFactory
     {
+        $this->authorize('create', Skill::class);
+
         if ($this->wantsTurboStream($this->request)) {
             $skill = $this->skills->build();
             if (($sess = $this->request->session()) && $sess->hasOldInput()) {
@@ -83,10 +97,12 @@ class SkillController extends AdminController
      *
      * @param Skill $skill
      * @return RedirectResponse|Response|ResponseFactory
-     * @throws BindingResolutionException
+     * @throws BindingResolutionException|AuthorizationException
      */
     public function edit(Skill $skill): RedirectResponse|Response|ResponseFactory
     {
+        $this->authorize('edit', $skill);
+
         if ($this->wantsTurboStream($this->request)) {
             if (($sess = $this->request->session()) && $sess->hasOldInput()) {
                 return $this->renderTurboStream('admin.skills.form.modal_stream', compact('skill'));
@@ -102,10 +118,10 @@ class SkillController extends AdminController
      *
      * Validate skill form and create skill, then redirect to skills index.
      *
-     * @param SkillStoreRequest $request
+     * @param SkillRequest $request
      * @return RedirectResponse
      */
-    public function store(SkillStoreRequest $request): RedirectResponse
+    public function store(SkillRequest $request): RedirectResponse
     {
         if ($this->skills->create($request->validated())) {
             return redirect()
@@ -127,11 +143,11 @@ class SkillController extends AdminController
      *
      * Validate skill form and update skill, then redirect to skills index.
      *
-     * @param SkillUpdateRequest $request
+     * @param SkillRequest $request
      * @param Skill $skill
      * @return RedirectResponse
      */
-    public function update(SkillUpdateRequest $request, Skill $skill): RedirectResponse
+    public function update(SkillRequest $request, Skill $skill): RedirectResponse
     {
         if ($this->skills->update($skill, $request->validated())) {
             return redirect()
@@ -155,9 +171,12 @@ class SkillController extends AdminController
      *
      * @param Skill $skill
      * @return Renderable|RedirectResponse
+     * @throws AuthorizationException
      */
     public function destroy(Skill $skill): Renderable|RedirectResponse
     {
+        $this->authorize('delete', $skill);
+
         $this->skills->delete($skill);
         return redirect()
             ->back()
